Refactor: MemberAccount관련 위치 변경 & 필터 조회 수정
- MemberAccount 중복 건 수정
- AccountList에 MemberAccount상태 변경 등록
- MemberAccount 필터 조회 수정

Refs: #19

+++ ordering/src/main/java/com/beyond/MKX/domain/account/member/dto/AmountRequest.java
@@ -12,6 +12,6 @@ import java.math.BigInteger;
 public class AmountRequest {
     @NotNull
     @Positive
-    private BigInteger amount;
+    private Long amount;
 }
